Add Chromium OS Developr Guide
The developer guide [1] is translated into a Markdown file.
Here's how the file is made:

- Run "turndown" tool [2] for preprocessing
- Fix miscellaneous errors manually

The text contents are kept as is except for very minor fixes
for punctuations, etc. There will be follow-up patches
to update the contents properly (ex. the instruction on
updating this document needs to be updated).

[1] http://dev.chromium.org/chromium-os/developer-guide
[2] http://domchristie.github.io/turndown/

BUG=None
TEST=Read the document with md_browser.py

Change-Id: I99d1c4d96da95c077a34e6a1c3c8912afcd1f617
Reviewed-on: https://chromium-review.googlesource.com/1073033
Reviewed-by: Satoru Takabayashi <satorux@chromium.org>
Reviewed-by: Daichi Hirono <hirono@chromium.org>
Commit-Queue: Satoru Takabayashi <satorux@chromium.org>
Tested-by: Satoru Takabayashi <satorux@chromium.org>