fuzzing.md: Re-order sections to remove repetition.
The document was not straightforward to read because a big chunk of the
instructions were repeated in the "platform package" and
"non-platform package" sections. This was problematic because any
guidelines on how to write a fuzzer would have to be forever
duplicated.

Fix that by extracting a "How to write a fuzzer" section with all the
common content, and keep the separate content in "How to build a
fuzzer" sections.

A follow-up CL will expand the "How to write a fuzzer" section.

BUG=None
TEST=Open in Markdown editor.

Change-Id: I4f5ceaa18d506a0e0a1c808c3427908515d06a11
Reviewed-on: https://chromium-review.googlesource.com/1079288
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>