Clarify process for flashing image to USB in docs
If auto-mount is enabled on a build computer, USB drives may sometimes
be corrupted even when flashing appears to have been successful. There
was a line in developer_guide.md, but it appeared after the instructions
to insert a USB drive and was easy to miss.

This commit moves that line earlier in the section so that the reader
will see that before inserting any USB drive.

Change-Id: Ibb3ece0502e92d638074862af9bb0de0ef2ef6fd
Reviewed-on: https://chromium-review.googlesource.com/1231394
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>