fuzzing: drop mention of `pkg-conf`
There is no such tool "pkg-conf" in the tree, and assuming this wanted
to tell people to use `pkg-config`, that's not something they should do
either -- that's the build tool that doesn't work for cross-compiling.

Lets drop the line entirely and just focus on GYP/GN builds.  If people
are using other projects, they should be able to fill in the gaps with
whatever build system is used.

Change-Id: Ibbfd2c329041507b206f1b3e747dc8433e7878a2
Reviewed-on: https://chromium-review.googlesource.com/1328264
Commit-Ready: Allen Webb <allenwebb@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jonathan Metzman <metzman@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>