sandboxing: suggest testing error paths for seccomp
This adds a short paragraph to the use of "strace" for generating
seccomp policies.  It reminds that it's easy to neglect syscalls
that are executed as a result of errors or other unexpected
conditions (for instance, abort()).  In such cases, the seccomp
failure can prevent full reporting of the error condition and
delay debugging.

BUG=none
TEST=none

Change-Id: Ida3818b50a15a423e4873e338c02e036ba039198
Reviewed-on: https://chromium-review.googlesource.com/1370685
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>