kernel_faq: Explicitly note that rejected kernel patches should be labeled CHROMIUM
The docs don't state what to do if a patch has been rejected or deemed
unsuitable for upstream. If we still need the patch, it should probably
be labeled CHROMIUM since while it's on the list it won't reach
mainline.

Labeling rejected patches as FROMLIST may give a false sense of security
to reviewers. This may be in part due to auto-reviewing scripts */me
ducks* which validate the backport does not differ from the post, but do
not surface the sentiment of the review feedback.

To avoid this situation (or at least to set expectations), we should
explicitly state that these patches should be labeled CHROMIUM when they
are backported.

Change-Id: I484baf18f8458a92fbf81660be8bf8bfa33bc387
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1669748
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>