kernel-faq: Mention cross-gdb and vmlinux in the debugging guide
For non-x86 devices, the host gdb may not work and so the proper gdb
coming from the corresponding cross-toolchain should be used. Actually,
even for x86-devices, there is no guarantee that the host and target
would be the same, so suggest to just always use the cross one.

In addition to that, make the doc mention simply loading vmlinux instead
of the specific object file.

BUG=none
TEST=Preview the rendered doc

Change-Id: Ifcaaf4e43552cff41e0bebb0dbdbbb8ac0bfd039
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/docs/+/1751162
Commit-Queue: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>