rust_on_cros: Update SLOT setting for first-party crates
Setting the SLOT to "${PV}/${PR}" is useful for third-party crates as it
allows different versions of the same crate to be installed
simultaneously.  However, there will only be one version of first-party
crates so using ${PV} for the main slot doesn't really make sense.  This
also causes problems when a developer does `cros_workon start` on a
first-party rust package as it causes file collision problems the first
time they try to emerge it.

Instead update the docs to recommend using "0" for the main SLOT
value (the subslot should still be ${PVR} to pick up changes) as this
better reflects the actual usage of the package.

BUG=none
TEST=none

Change-Id: Ic7c785d1d9af53be290140ef3b32dc04c2f0a721
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/docs/+/1824577
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>
Auto-Submit: Chirantan Ekbote <chirantan@chromium.org>