testing/fuzzing.md: Hint about "noclean" for line by line coverage
Add into the documentation a hint about how to get the per-source-file
HTML files in the cross_fuzz coverage report.

The tip is to use FEATURES="noclean" for building each of the packages
that contain the related source files, since apparently without it the
llvm-cov command is unable to generate the HTML files that allow to
browse the sources together with the coverage information.

BUG=none

Change-Id: I1a1dd35a4d6955b1c976a75b09fa585a175b4d84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/docs/+/1880036
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Auto-Submit: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>