firmware_test_manual: create new doc
This relocates the "firmware test manual" to public Chromium OS
documentation. There are some minor edits for style and clarification.

In addition the following sections have been removed, as they referred
to internal details:
  - "Number of Released RO"
  - "Device Idiosyncracies"
  - "Miscellaneous"

Updated filenames to diagrams added to the images/ directory.

This is in anticipation of adding a section documenting a process for
debugging boot failure, which we would like to share publicly.

BUG=None
TEST=None

Change-Id: I4319eac2c9c468f396f423d15802543dde0c5a45
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/docs/+/1972405
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>