work_on_branch: Remove wrong instruction about Change-Id.
The instruction to recreate Change-Id when cherry-picking to
a (release or otherwise) branch is a vestige from the era when
Gerrit didn't allow the same Change-Id across branches. Now it does,
and referencing Change-Id is a useful way of tracking that change
across all branches. So this line is actually harmful, remove it.

BUG=None
TEST=None

Change-Id: I7dc1711f1e1fbd00dcbc1b63f128616cd89268ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/docs/+/1986206
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Commit-Queue: Kirtika Ruchandani <kirtika@chromium.org>