contributing: Cq-Depend can't work with Change-Id
As per an email from Andrii Shyshkalov on April 8th 2019:

> Sadly, Change-Id isn't supported with new footers. We were aware of
> use case you described, but because Change-Id isn't unique even when
> narrowed to a single ref of a single repo (see http://b/68861593),
> resolution of Change-Id to an Change number (actually unique) may
> change with time.  Although it's likely possible to work around
> this, as I was told this isn't a critical use case for CrOS, so it
> shouldn't be blocking parallel CQ project yet may be added later if
> necessary.

As far as I'm aware it still hasn't been added.  However, the docs
written April 13th 2019 (according to git logs) say that it's
supported.

I'm going to assume that this is a mistake in the docs.  I don't see
anyone using the Change-Id.

NOTE: if the docs are correct, it would be good to add an example
since it's unclear if the syntax would be chromium:change-id or just a
raw changei-id.

BUG=None
TEST=None

Change-Id: I99cf9d94fc1bae7f5b3d13ad87a9b206ae50502b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/docs/+/2154236
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>