security_severity_guidelines: Update guidelines for critical bugs.
The context of critical bugs, especially in the browser, has changed.
Update the document to reflect that.

Change the guidelines from "time to fix deployed to all users", which
is hard for us to control (because of pinned versions and the like),
to "time to fix available on the stable channel", which is something
we can more easily control and seems to be the metric large
organizations are tracking.

Also freshen up the High-severity section to clarify the types of
bypasses.

BUG=None
TEST=Open in Markdown viewer.

Change-Id: Ibf6cd73b03439c0c3b3fe2c179f47434a717c19f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/docs/+/2177333
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>