simple_chrome_workflow.md: Change suggestion for turning DCHECKs on.
In my experience, the recommendation of turning DCHECKs on is
problematic for many developers, especially newcomers. I think we
should highlight that DCHECKs most likely will result is random crashes
and that it's a choice to turn them on to help development collectively.

See related bugs:
- crbug.com/913750
- crbug.com/905882

Change-Id: Ie1b55a62c2b918a03a12fcb580bc56f5e202e910
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/docs/+/2178471
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Achuith Bhandarkar <achuith@chromium.org>
Tested-by: Achuith Bhandarkar <achuith@chromium.org>