rust_on_cros: clarify how to specify dependencies
For the example ebuild for a third-party crate, add more examples of
dependencies with neither '^' nor '~' specified. These are handled the
same as dependencies with '^', but a brief survey of third party ebuilds
in dev-rust shows that this isn't as clear as it could be.

Change-Id: Iaf9436e3f74cfd51c03525f5dff7f793784bf34d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/docs/+/2308789
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>