developer_guide: Clarity changes for the "Install git and curl" section
Instructions in the developer guide, are typically structured with the
rationale/conditional of an instruction first and then the instruction/s
to run are in a following segment of code. The "Install git and curl"
section has important comments/conditionals in the code section, which
makes it easier for readers to miss them. This can be especially
detrimental for Googlers, as they may unwittingly adjust their Python
config and put their workstation out of sync.

Change-Id: Ief63f6f1595955dabd3ff036bcb92c9fd513f159
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/docs/+/2425971
Commit-Queue: Daniel Ng <danielng@google.com>
Tested-by: Daniel Ng <danielng@google.com>
Reviewed-by: Stephane Belmon <sbelmon@google.com>