ebuild_faq: Add virtual version policy for chromeos-overlay
Previously this overlay did not have an entry and people seem to have
been using `1.5` as a default version for virtuals in chromeos-overlay,
but that conflicts with the version used by "project-" overlays. Add a
new entry to reflect that virtuals in chromeos-overlay should override
those in chromiumos-overlay but not anything else.

BUG=None
TEST=None

Change-Id: I6e733d1a71cfa3fdf274e3fb93b0eac88d469f9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/docs/+/2506132
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Auto-Submit: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Stephane Belmon <sbelmon@google.com>
Commit-Queue: Stephane Belmon <sbelmon@google.com>